Splits cli add tests on different files
