+++ tests/cli/add/test_cli_add.py
@@ -1,9 +1,7 @@
-"""Tests for adding via CLI"""
+"""Tests for adding tags via CLI"""
 import logging
 
-from housekeeper.cli import add
-
-# tests adding tags
+from housekeeper.cli.add import tag_cmd
 
 
 def test_add_two_tags(populated_context, cli_runner, caplog):
@@ -14,7 +12,7 @@ def test_add_two_tags(populated_context, cli_runner, caplog):
     tag1 = "new-tag"
     tag2 = "other-tag"
     # WHEN trying to add two tags to the an existing file
-    result = cli_runner.invoke(add.tag, [tag1, tag2], obj=populated_context)
+    result = cli_runner.invoke(tag_cmd, [tag1, tag2], obj=populated_context)
     # THEN assert it has a zero exit status
     assert result.exit_code == 0
     # THEN check that the tags are logged
@@ -36,7 +34,7 @@ def test_add_existing_tag_existing_file(populated_context, cli_runner, caplog):
 
     # WHEN trying to add the existing tag to the file
     result = cli_runner.invoke(
-        add.tag, [tag, "-f", str(file_id)], obj=populated_context
+        tag_cmd, [tag, "-f", str(file_id)], obj=populated_context
     )
     # THEN assert it has a non zero exit status
     assert result.exit_code == 0
@@ -58,7 +56,7 @@ def test_add_tag_existing_file(populated_context, cli_runner, caplog):
 
     # WHEN trying to add a tag to the existing file
     result = cli_runner.invoke(
-        add.tag, [tag, "-f", str(file_id)], obj=populated_context
+        tag_cmd, [tag, "-f", str(file_id)], obj=populated_context
     )
     # THEN assert it has a zero exit status
     assert result.exit_code == 0
@@ -78,87 +76,9 @@ def test_add_tag_non_existing_file(populated_context, cli_runner, caplog):
 
     # WHEN trying to add a tag to the non existing file
     result = cli_runner.invoke(
-        add.tag, ["new-tag", "-f", str(missing_file_id)], obj=populated_context
+        tag_cmd, ["new-tag", "-f", str(missing_file_id)], obj=populated_context
     )
     # THEN assert it has a non zero exit status
     assert result.exit_code == 1
     # THEN check that the error message is displayed
     assert "unable to find file" in caplog.text
-
-
-# tests adding bundles
-
-
-def test_add_existing_bundle(populated_context, cli_runner, caplog):
-    """Test to add a bundle that exists"""
-    caplog.set_level(logging.DEBUG)
-    # GIVEN a context with a populated store and a cli runner
-    store = populated_context["store"]
-    # GIVEN a existing bundle
-    bundle_obj = store.Bundle.query.first()
-    assert bundle_obj
-    bundle_name = bundle_obj.name
-
-    # WHEN trying to add a bundle that exists
-    result = cli_runner.invoke(add.bundle, [bundle_name], obj=populated_context)
-
-    # THEN assert it has a non zero exit status
-    assert result.exit_code == 1
-    # THEN check that the error message is displayed
-    assert "bundle name already exists" in caplog.text
-
-
-def test_add_bundle(base_context, cli_runner, case_id, caplog):
-    """Test to add a new bundle"""
-    caplog.set_level(logging.DEBUG)
-    # GIVEN a context with a empty store and a cli runner
-    bundle_name = case_id
-
-    # WHEN trying to add a bundle
-    result = cli_runner.invoke(add.bundle, [bundle_name], obj=base_context)
-
-    # THEN assert it succeded
-    assert result.exit_code == 0
-    # THEN check that the proper information is displayed
-    assert "new bundle added" in caplog.text
-
-
-# tests adding file to bundle
-
-
-def test_add_file_non_existing_bundle(
-    base_context, cli_runner, case_id, second_sample_vcf, caplog
-):
-    """Test to add a file to a non existing bundle"""
-    caplog.set_level(logging.DEBUG)
-    # GIVEN a context with a empty store and a cli runner
-    unknown_bundle_name = case_id
-
-    # WHEN trying to add a bundle
-    result = cli_runner.invoke(
-        add.file_cmd, [unknown_bundle_name, str(second_sample_vcf)], obj=base_context
-    )
-
-    # THEN assert it fails
-    assert result.exit_code == 1
-    # THEN check that the proper information is displayed
-    assert f"unknown bundle: {unknown_bundle_name}" in caplog.text
-
-
-def test_add_file_existing_bundle(
-    populated_context, cli_runner, case_id, second_sample_vcf, caplog
-):
-    """Test to add a file to a existing bundle"""
-    caplog.set_level(logging.DEBUG)
-    # GIVEN a context with a populated store and a cli runner
-    bundle_name = case_id
-
-    # WHEN trying to add the file to a bundle
-    result = cli_runner.invoke(
-        add.file_cmd, [bundle_name, str(second_sample_vcf)], obj=populated_context
-    )
-
-    # THEN assert it succedes
-    assert result.exit_code == 0
-    # THEN check that the proper information is displayed
-    assert "new file added" in caplog.text
